+++ CHANGELOG.md
@@ -17,6 +17,11 @@ that includes `constructor` key to override the class constructor.
 * Three inner (private) attributes (`formName`, `selectBoxName`, and `fieldNamePrefix`) have
 been made public.
 
+### `Cape.DataStore`
+
+* Allow the constructor of `Cape.DataStore` to take option values.
+* Allow `Cape.DataStore.create()` to accept options values for constructor.
+
 ### `Cape.ResourceAgent`
 
 * Add a new class called `Cape.ResourceAgent`, which performs CRUD operations